+++ templates/product_page.html
@@ -26,7 +26,6 @@
 
 <div class="prod_page_card">
         <table class="product_card_prod_page">
-
              <tr>
             <td class="table_header">Founded</td><td>{{ brand.founded }}</td>
          </tr>
@@ -35,6 +34,12 @@
          </tr>
              <tr>
             <td class="table_header">Products</td><td>{{ brand.products }}</td>
+         </tr>
+      <tr>
+            <td class="table_header">Props</td><td>{{ brand.Props}}</td>
+         </tr>
+      <tr>
+            <td class="table_header">Books</td><td>{{ brand.Books}}</td>
          </tr>
         </table>
 </div>
